requirements.txt
# esse arquivo aqui vai ser usado pelo heroku,
# usem o requirements_local.txt pra desenvolver localmente!

Django==1.5.3
django-mptt==0.6.0
django-model-utils==1.1.0
django-annoying==0.7.7

# heroku
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
static==0.4
